move files + add storybook for auth screen
