+++ src/stories/AuthScreen.stories.js
@@ -1,0 +1,12 @@
+  
+import React from 'react';
+import AuthScreen from '../screens/AuthScreen';
+
+export default {
+  title: 'AuthScreen',
+  component: AuthScreen,
+};
+
+const Template = args => <AuthScreen {...args} />;
+
+export const Default = Template.bind({});
